requirements.txt
django==2.0.4
djangorestframework==3.8.2
markdown==2.6.11
django-filter==1.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
django-rest-swagger==2.2.0
pillow
